Split ToolSpec into a separate file and add ToolVersion
